refactor: type 분리 및 import 정리
