+++ components/src/types/system.ts
@@ -1,0 +1,3 @@
+export type ModalPosition = 'center' | 'bottom';
+
+export type Size = 'small' | 'medium' | 'large';
